UUIDgenerate {uuid}	R Documentation

UUID Functions

Description

UUIDgenerate generates new Universally Unique Identifiers. It can be either time-based or random.

UUIDparse parses one of more UUIDs it string form and converts them to other internal formats.

UUIDvalidate checks the valitiy of UUIDs in string form.

Usage

UUIDgenerate(use.time = NA, n = 1L, output = c("string", "raw", "uuid"))
UUIDparse(what, output = c("uuid", "string", "raw", "logical"))
UUIDvalidate(what)

Arguments

use.time
logical, if TRUE then time-based UUID is generated, if FALSE then a random UUID is generated, if NA then random one is generated if a sufficiently reliable source of random numbers can be found, otherwise a time-based UUID is generated.

n
integer, number of UUIDs to generate.

output
type of the output. Valid types are: "string" for a character vector with UUIDs in textual representation (always lowercase), "raw" for a vector or matrix of raw bytes, "uuid" for an object of the class UUID and "logical" which only reports failure/success of the parsing, but not the actual values.

what
character vector which will be parsed into UUIDs.

Value

UUIDgenerate and UUIDparse values depend on the output argument as follows:

"string"
character vector with each element UUID in lowercase form, for UUIDparse strings that cannot be parsed will result in NA values

"raw"
raw vector with the UUIDs stores each as 16 bytes seqeuntially. If the output is more than one UUID then the result is a raw matrix with 16 rows and as many columns as there are input elements.

"uuid"
object of the class UUID which is a vector of UUIDs in 128-bit internal representation.

"logical"
only supported in UUIDparse and return code TRUE for valid UUID, FALSE for invalid input and NA for NA input.

UUIDvalidate is just a shorthand for UUIDparse(what, output="logical").

Note

The first argument is not n for historical reasons, beause the first version did only generate a single UUID.

Author(s)

Simon Urbanek, based on libuuid by Theodore Ts'o.

Examples

UUIDgenerate()
UUIDgenerate(TRUE)
UUIDgenerate(FALSE)

## see if the randomness is any good
length(unique(UUIDgenerate(n=1000)))

## generate a native UUID vector
(u <- UUIDgenerate(n=3, output="uuid"))

as.character(u)
as.raw(u[1])

UUIDgenerate(output="raw")

## see ?UUID for more examples on UUID objects

[Package uuid version 1.1-0 Index]